requirements.txt
numpy
scipy>=1.10

# Piper TTS
soundfile>=0.12
piper-tts>=1.2

resampy
webrtcvad
faster-whisper
pytest
requests

# ElevenLabs TTS
elevenlabs

faiss-cpu
sentence-transformers
watchfiles

discord.py[voice]==2.3.2

# Optional: phrase models and MIDI support
# CPU builds by default from PyPI. For NVIDIA GPU builds see
# requirements-gpu-cu121.txt or requirements-gpu-cu118.txt.
torch>=2.5
torchaudio>=2.5  # for decoding and saving audio
onnxruntime
mido
transformers>=4.45
accelerate>=0.33
safetensors>=0.4.2

# --- Riffusion stack (pinned) ---
# Pinned versions used by the Riffusion feature to avoid drift.
# These do not alter existing workflows but ensure stable builds for Riffusion.
diffusers==0.31.0
torch==2.6.0+cu124
transformers==4.56.1
accelerate==1.10.1
pillow==11.3.0
librosa==0.11.0
soundfile==0.13.1
numpy==2.2.6
scipy==1.15.3
